requirements.txt
appdirs==1.4.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
configobj==5.0.6
cryptography==2.8
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
future==0.18.2
idna==2.9
lxml==4.5.0
MechanicalSoup==0.12.0
parse==1.15.0
peewee==3.13.2
pycparser==2.20
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
python-telegram-bot==12.5.1
requests==2.23.0
requests-html==0.10.0
six==1.14.0
soupsieve==2.0
tornado==6.0.4
tqdm==4.44.1
twitter-scraper==0.4.1
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
